src/ChapterNine/CommissionEmployee.java
package ChapterNine;

public class CommissionEmployee {
    private final String firstName;
    private final String lastName;
    private final String socialSecurityNumbers;
    private double grossSales;
    private double commissionRate;

    public CommissionEmployee(String firstName, String lastName, String socialSecurityNumbers, double grossSales, double commissionRate) {

        if (grossSales < 0.0) {
            throw new IllegalArgumentException("Gross sales must be >= 0.0");
        }

        if (commissionRate <= 0.0 || commissionRate >= 1.0) {
            throw new IllegalArgumentException("Commission rate must be > 0.0 and < 1.0");
        }
        this.firstName = firstName;
        this.lastName = lastName;
        this.commissionRate = commissionRate;
        this.socialSecurityNumbers = socialSecurityNumbers;
        this.grossSales = grossSales;
    }

    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public String getSocialSecurityNumbers() {
        return socialSecurityNumbers;
    }

    public double getGrossSales() {
        return grossSales;
    }

    public double getCommissionRate() {
        return commissionRate;
    }

    public void setGrossSales(double grossSales) {
        if (grossSales < 0.0) throw new IllegalArgumentException("Gross sales must be >= 0.0");
        else this.grossSales = grossSales;
    }

    public void setCommissionRate(double commissionRate) {
        if (commissionRate <= 0.0 || commissionRate >= 1.0)
            throw new IllegalArgumentException("Commission rate must be > 0.0 and < 1.0");
        else this.commissionRate = commissionRate;
    }


    public double earnings() {
        return (commissionRate * grossSales);

    }
}

